refactor(#223): PaymebtModal을 cars/components/ 로 이동, 사용하지 않는 Payment 파일 삭제
